// src/types/alarm-status.enum.ts
export enum AlarmStatus {
    ALARM = "LE",
    ARMED = "AT",
    DISARMED = "DA",
    PARTIALLY = "AP",
    POWER_FAILURE = "FC",
    POWER_RESTORED = "RFC",
    IMAGE = "IM",
    ERROR_DISARMED = "EDA",
    ERROR_ARMED_TOTAL = "EAT",
    ERROR_PARTIALLY = "EAP",
    ERROR_ARMED_TOTAL_COMMUNICATIONS = "EAT-COM",
    ERROR_DISARMED_COMMUNICATIONS = "EDA-COM",
    ERROR_PARTIALLY_COMMUNICATIONS = "EAP-COM",
    GENERAL_ERROR = "ERROR",
}